% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{file_iconv}
\alias{file_iconv}
\title{Converts encoding of characters}
\usage{
file_iconv(path, from, to = "UTF-8")
}
\arguments{
\item{path}{file path}

\item{from}{the code set in which the input
is encoded.}

\item{to}{the code set to which the output is
to be converted.}
}
\description{
Read a file, convert the encoding
of characters and print the result.
}
\examples{
d_ <- system.file(package = "fpeek", "datafiles")

file_head_show(file.path(d_, "cigale-ISO-8859-1.txt"))
file_iconv(file.path(d_, "cigale-ISO-8859-1.txt"),
  from = "ISO-8859-1", to = "UTF-8")

file_head_show(file.path(d_, "Windows_1252.txt"))
file_iconv(file.path(d_, "Windows_1252.txt"),
           from = "WINDOWS-1252", to = "UTF-8")

file_head_show(file.path(d_, "UTF16_BE.txt"))
file_iconv(file.path(d_, "UTF16_BE.txt"),
           from = "UTF-16", to = "UTF-8")
}
