% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{file_nlines}
\alias{file_nlines}
\title{number of lines of a file}
\usage{
file_nlines(path)
}
\arguments{
\item{path}{file path}
}
\value{
number of lines as an integer
}
\description{
return the number of lines
found in a file.
}
\examples{
f <- system.file(package = "fpeek",
  "datafiles", "cigale-ISO-8859-1.txt")
file_nlines(f)
}
