% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{file_grep_show}
\alias{file_grep_show}
\alias{grep}
\alias{file_grep}
\title{grep a file and print the results}
\usage{
file_grep_show(path, regexp)
}
\arguments{
\item{path}{file path}

\item{regexp}{regular expression as a string}
}
\description{
print lines
of a file matching a regular expression.
}
\examples{
f <- system.file(package = "fpeek",
                 "datafiles", "cigale-ISO-8859-1.txt")
file_grep_show(f, regexp = "([C|c]igale|fut)")
file_grep_show(f, regexp = "[^i]er")
}
