% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{file_head_show}
\alias{file_head_show}
\alias{head}
\alias{file_head}
\alias{file_head_char}
\title{print the first lines of files}
\usage{
file_head_show(path, n = 10)

file_head_char(path, n = 10)
}
\arguments{
\item{path}{file path}

\item{n}{number of lines to print}
}
\description{
print the first \code{n} lines
of a file.
}
\examples{
f <- system.file(package = "fpeek",
  "datafiles", "cigale-ISO-8859-1.txt")
file_head_show(f, n = 4)
file_head_char(f, n = 4)
}
