% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{file_iconv_show}
\alias{file_iconv_show}
\alias{iconv}
\alias{file_iconv}
\alias{file_iconv_write}
\title{Converts encoding of characters}
\usage{
file_iconv_show(path, from, to = "UTF-8")

file_iconv_write(path, newfile, from, to = "UTF-8")
}
\arguments{
\item{path}{file path}

\item{from}{the code set in which the input
is encoded.}

\item{to}{the code set to which the output is
to be converted.}

\item{newfile}{result file}
}
\description{
Read a file, convert the encoding
of characters and print the result.
}
\examples{
la_cigale <- system.file(package = "fpeek", "datafiles",
  "cigale-ISO-8859-1.txt")

file_head_show(la_cigale)
file_iconv_show(la_cigale, from = "ISO-8859-1", to = "UTF-8")

file <- system.file(package = "fpeek", "datafiles",
  "Windows_1252.txt")
file_head_show(file)
file_iconv_show(file, from = "WINDOWS-1252", to = "UTF-8")

file <- system.file(package = "fpeek", "datafiles",
  "UTF16_BE.txt")
file_head_show(file)
file_iconv_show(file, from = "UTF-16", to = "UTF-8")
la_cigale <- system.file(package = "fpeek", "datafiles",
  "cigale-ISO-8859-1.txt")

file_utf8 <- tempfile()
file_iconv_write(path = la_cigale, newfile = file_utf8,
  from = "WINDOWS-1252", to = "UTF-8")

}
