% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{less_show}
\alias{less_show}
\alias{less}
\alias{file_less}
\title{print the last lines of files}
\usage{
less_show(path, n = 6)
}
\arguments{
\item{path}{file path}

\item{n}{number of lines to print}
}
\description{
print the last \code{n} lines
of a file.
}
\section{commands}{


type \code{q} to quit.

type any key to see next lines.

type \code{p} to see previous lines (not implemented yet).
}

\examples{
if( interactive() ){
  f <- system.file(package = "fpeek",
    "datafiles", "cigale-ISO-8859-1.txt")
  less_show(f)
}
}
