% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{peek_tail}
\alias{peek_tail}
\title{print the last lines of files}
\usage{
peek_tail(path, n = 10, intern = FALSE)
}
\arguments{
\item{path}{file path}

\item{n}{number of lines to print}

\item{intern}{a logical which indicates whether to capture
the output as an R character vector or to print the output
in the R console.}
}
\description{
print the last \code{n} lines
of a file.
}
\examples{
f <- system.file(package = "fpeek",
  "datafiles", "cigfou-ISO-8859-1.txt")
peek_tail(f, n = 4)
peek_tail(f, n = 4, intern = TRUE)
}
