% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{file_tail_show}
\alias{file_tail_show}
\alias{tail}
\alias{file_tail}
\alias{file_tail_char}
\title{print the last lines of files}
\usage{
file_tail_show(path, n = 10)

file_tail_char(path, n = 10)
}
\arguments{
\item{path}{file path}

\item{n}{number of lines to print}
}
\description{
print the last \code{n} lines
of a file.
}
\examples{
f <- system.file(package = "fpeek",
  "datafiles", "cigale-ISO-8859-1.txt")
file_tail_show(f, n = 4)
file_tail_char(f, n = 4)
}
