% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{wc_l}
\alias{wc_l}
\title{number of lines of a file}
\usage{
wc_l(path)
}
\arguments{
\item{path}{file path}
}
\value{
number of lines as an integer
}
\description{
return the number of lines
found in a file.
}
\examples{
f <- system.file(package = "fpeek",
  "datafiles", "cigale-ISO-8859-1.txt")
wc_l(f)
}
